SCHOOL/senior_project/backend/requirements.txt
# OpenAI
openai>=1.20.0

# Web framework and WebSockets
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
python-engineio>=4.4.0
eventlet>=0.33.0
hypercorn>=0.14.0

# Audio processing
numpy>=1.22.0
sounddevice>=0.4.6
soundfile>=0.12.1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
boto3>=1.26.0  # For AWS Polly if needed

# WSGI Server
gunicorn>=21.0.0
websockets>=11.0.2 